requirements.txt
google-generativeai>=0.8.0
PyQt6>=6.6.0
